# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a mycrs server.",
        "title": "Swagger MYCRS API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/mycrs",
    "paths": {
...
        "/dev/requestRevoke/{requestCode}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dev"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request Code",
                        "name": "requestCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/dev/resetPin/{resetCode}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dev"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reset Code",
                        "name": "resetCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}